Ran through CppCheck

+++ SFMLSpaceGame/EntityFactory.cpp
@@ -195,7 +195,7 @@ void EntityFactory::MakeIntoShip(EntityHandle& ent, const std::string& shipName,
 	auto origin = sf::Vector2f(spriteBox.width, spriteBox.height) / 2.f;
 	auto& verts = shipStats->GetColliderVertices();
 	auto shape = sf::ConvexShape(verts.size());
-	for (int i = 0; i < verts.size(); i++)
+	for (size_t i = 0; i < verts.size(); i++)
 	{
 		shape.setPoint(i, (verts[i] * METERS_PER_PIXEL) - origin);
 	}
